Refactor core library
`RelativityParser` is now a static class that defines properties for invariant culture parsers in both the local machine time zone and UTC, and a `Current` property that uses `AsyncLocal<T>` to store the Relativity parser for the current asychronous control flow. This means that `RelativityParser.Current` can be set at the start of an HTTP request and re-used throughout the request as required.

The functionality of a parser has been abstracted into the `IRelativityParser` interface. The `IRelativityParserFactory` interface defines the methods for registering and retrieving parser instances. `RelativityParserFactory` is the default implementation of this interface, and also defines a static `Default` property that can be used to get an instance of the factory in scenarios where dependency injection is not being used.

Parsers now have an associated `TimeZoneInfo` instead of this being passed into the methods defined on the `IRelativityParser` interface. `RelativityParserFactory` has built-in logic for cloning existing parsers if the time zone specified when requesting a parser differs from the time zone of the parser.