Relativity V2 implementation (#1)
* Update build to use Cake

* Refactor core library

`RelativityParser` is now a static class that defines properties for invariant culture parsers in both the local machine time zone and UTC, and a `Current` property that uses `AsyncLocal<T>` to store the Relativity parser for the current asychronous control flow. This means that `RelativityParser.Current` can be set at the start of an HTTP request and re-used throughout the request as required.

The functionality of a parser has been abstracted into the `IRelativityParser` interface. The `IRelativityParserFactory` interface defines the methods for registering and retrieving parser instances. `RelativityParserFactory` is the default implementation of this interface, and also defines a static `Default` property that can be used to get an instance of the factory in scenarios where dependency injection is not being used.

Parsers now have an associated `TimeZoneInfo` instead of this being passed into the methods defined on the `IRelativityParser` interface. `RelativityParserFactory` has built-in logic for cloning existing parsers if the time zone specified when requesting a parser differs from the time zone of the parser.

* Dependency injection support

The IntelligentPlant.Relativity.DependencyInjection project contains methods for registering a singleton `IRelativityParserFactory` with the DI container, and for registering additional parser configurations not specified in the well-known parser definitions.

* Add ASP.NET Core middleware

IntelligentPlant.Relativity.AspNetCore defines middleware and extensions that can be used to set `RelativityParser.Current` for an HTTP request based on the calling user's culture and time zone.

Time zone information can be supplied via query string, cookie, request header, user claim, or a custom mechanism. If time zone information is not specified, UTC will be assumed.

* Add OWIN middleware

IntelligentPlant.Relativity.Owin defines middleware for setting `RelativityParser.Current` for an incoming HTTP request.

The time zone for the request can be set via query string, cookie, request header, user claim, or a custom mechanism. If time zone information is not specified, UTC will be assumed.

* Update tests

* Update README.md

* Update examples