Add simple cache to RegexHelper
RegexHelper is now resonsible for creating Regex instances instead of just the regex patterns. It maintains an internal cache (ConcurrentDictionary<string, Regex>) of regex instances that it has created and reuses the instances where possible.

The cache key is derived from the parser keywords for the pattern and the decimal separator for the culture that the pattern is being created for. This allows us to create e.g. a relative timestamp pattern that uses invariant keywords but uses a comma as a decimal separator for non-English cultures.