Always use requested culture when returning a parser (#11)
* Always use requested parser culture

When a parser is requested for a culture that does not have a registered parser configuration, return a parser that uses the base time and time offset keywords for the invariant parser but uses the requested culture for parsing absolute timestamps and relative offset quantities.

* Add simple cache to RegexHelper

RegexHelper is now resonsible for creating Regex instances instead of just the regex patterns. It maintains an internal cache (ConcurrentDictionary<string, Regex>) of regex instances that it has created and reuses the instances where possible.

The cache key is derived from the parser keywords for the pattern and the decimal separator for the culture that the pattern is being created for. This allows us to create e.g. a relative timestamp pattern that uses invariant keywords but uses a comma as a decimal separator for non-English cultures.

* Cache parser configurations, not parser instances

RelativityParserFactory now caches RelativityParserConfiguration objects rather than IRelativityParser instances.

* Always verify parser culture in unit tests