Convert to duration with custom decimal places
Modifies `RelativityParserExtensions.ConvertToDuration` so that instead of a simple truncate yes/no option, it allows the number of decimal places to be specified and optionally allows the unit to use for the duration to be provided instead of being inferred from the magnitude of the TimeSpan.

This is a breaking change.