Update parser extensions
The `ConvertToDuration` extension method for `IRelativityParser` has been refactored into two overloads. The first takes only the `TimeSpan` to convert and creates a duration string that uses composite units. The second takes the units to define the duration string in and optionally the number of decimal places to round the magnitude to. Rounding is always performed away from zero e.g. specifying one decimal place and a magnitude of 1.12 will round the value to 1.2 instead of down to 1.1.

The commit also adds various `GetSymbol` extension method overloads, which allow the caller to retrieve the symbol defined by the parser for a given relative time origin, relative time offset unit, or duration unit.