Composite unit durations (#19)
* Allow composite durations to be specified

Duration strings and relative time offsets can now be specified using composite units. For example, instead of expressing a 90 second duration as `90S` or `1.5M`, it is now possible to specify `1M 30S`. White space between the duration components is ignored.

* Add enums to represent duration/offset unit types

* Update parser extensions

The `ConvertToDuration` extension method for `IRelativityParser` has been refactored into two overloads. The first takes only the `TimeSpan` to convert and creates a duration string that uses composite units. The second takes the units to define the duration string in and optionally the number of decimal places to round the magnitude to. Rounding is always performed away from zero e.g. specifying one decimal place and a magnitude of 1.12 will round the value to 1.2 instead of down to 1.1.

The commit also adds various `GetSymbol` extension method overloads, which allow the caller to retrieve the symbol defined by the parser for a given relative time origin, relative time offset unit, or duration unit.

* Add TimeSpan extensions

Adds `ToDurationString` extensions to `TimeSpan` that hook into the `ConvertToDuration` extensions for `IRelativityParser`.

* Add DurationBuilder

`DurationBuilder` is a new builder type to simplify the creation of duration strings.

* Update unit tests

* Update README files

Adds information about composite units for relative times and durations.

* Update version.json